// backend/src/modules/member/core/events/MemberDeleted.d.ts
import DomainEvent from '../../../shared/core/events/DomainEvent';
import type DateTime from '../../../shared/core/objects/DateTime';
import type IdEntity from '../../../shared/core/objects/IdEntity';
export default class MemberDeleted extends DomainEvent {
    constructor(key: string, date: DateTime, actor: IdEntity, idProject: IdEntity, idEntity: IdEntity);
}
//# sourceMappingURL=MemberDeleted.d.ts.map